+++ source/nmbd/nmbd.c
@@ -381,18 +381,18 @@ static void msg_nmbd_send_packet(struct messaging_context *msg,
 	const struct sockaddr_storage *pss;
 	const struct in_addr *local_ip;
 
-	DEBUG(10, ("Received send_packet from %d\n", procid_to_pid(&src)));
+	DEBUG(10, ("Received send_packet from %u\n", (unsigned int)procid_to_pid(&src)));
 
 	if (data->length != sizeof(struct packet_struct)) {
-		DEBUG(2, ("Discarding invalid packet length from %d\n",
-			  procid_to_pid(&src)));
+		DEBUG(2, ("Discarding invalid packet length from %u\n",
+			  (unsigned int)procid_to_pid(&src)));
 		return;
 	}
 
 	if ((p->packet_type != NMB_PACKET) &&
 	    (p->packet_type != DGRAM_PACKET)) {
-		DEBUG(2, ("Discarding invalid packet type from %d: %d\n",
-			  procid_to_pid(&src), p->packet_type));
+		DEBUG(2, ("Discarding invalid packet type from %u: %d\n",
+			  (unsigned int)procid_to_pid(&src), p->packet_type));
 		return;
 	}
 
@@ -400,8 +400,8 @@ static void msg_nmbd_send_packet(struct messaging_context *msg,
 	pss = iface_ip(&ss);
 
 	if (pss == NULL) {
-		DEBUG(2, ("Could not find ip for packet from %d\n",
-			  procid_to_pid(&src)));
+		DEBUG(2, ("Could not find ip for packet from %u\n",
+			  (unsigned int)procid_to_pid(&src)));
 		return;
 	}
