+++ source/winbindd/winbindd_cache.c
@@ -2624,9 +2624,9 @@ void cache_store_response(pid_t pid, struct winbindd_response *response)
 		return;
 
 	DEBUG(10, ("Storing response for pid %d, len %d\n",
-		   pid, response->length));
+		   (int)pid, response->length));
 
-	fstr_sprintf(key_str, "DR/%d", pid);
+	fstr_sprintf(key_str, "DR/%d", (int)pid);
 	if (tdb_store(wcache->tdb, string_tdb_data(key_str), 
 		      make_tdb_data((uint8 *)response, sizeof(*response)),
 		      TDB_REPLACE) == -1)
@@ -2640,7 +2640,7 @@ void cache_store_response(pid_t pid, struct winbindd_response *response)
 	DEBUG(10, ("Storing extra data: len=%d\n",
 		   (int)(response->length - sizeof(*response))));
 
-	fstr_sprintf(key_str, "DE/%d", pid);
+	fstr_sprintf(key_str, "DE/%d", (int)pid);
 	if (tdb_store(wcache->tdb, string_tdb_data(key_str),
 		      make_tdb_data((uint8 *)response->extra_data.data,
 				    response->length - sizeof(*response)),
@@ -2650,7 +2650,7 @@ void cache_store_response(pid_t pid, struct winbindd_response *response)
 	/* We could not store the extra data, make sure the tdb does not
 	 * contain a main record with wrong dangling extra data */
 
-	fstr_sprintf(key_str, "DR/%d", pid);
+	fstr_sprintf(key_str, "DR/%d", (int)pid);
 	tdb_delete(wcache->tdb, string_tdb_data(key_str));
 
 	return;
@@ -2664,9 +2664,9 @@ bool cache_retrieve_response(pid_t pid, struct winbindd_response * response)
 	if (!init_wcache())
 		return false;
 
-	DEBUG(10, ("Retrieving response for pid %d\n", pid));
+	DEBUG(10, ("Retrieving response for pid %d\n", (int)pid));
 
-	fstr_sprintf(key_str, "DR/%d", pid);
+	fstr_sprintf(key_str, "DR/%d", (int)pid);
 	data = tdb_fetch(wcache->tdb, string_tdb_data(key_str));
 
 	if (data.dptr == NULL)
@@ -2688,7 +2688,7 @@ bool cache_retrieve_response(pid_t pid, struct winbindd_response * response)
 	DEBUG(10, ("Retrieving extra data length=%d\n",
 		   (int)(response->length - sizeof(*response))));
 
-	fstr_sprintf(key_str, "DE/%d", pid);
+	fstr_sprintf(key_str, "DE/%d", (int)pid);
 	data = tdb_fetch(wcache->tdb, string_tdb_data(key_str));
 
 	if (data.dptr == NULL) {
@@ -2715,10 +2715,10 @@ void cache_cleanup_response(pid_t pid)
 	if (!init_wcache())
 		return;
 
-	fstr_sprintf(key_str, "DR/%d", pid);
+	fstr_sprintf(key_str, "DR/%d", (int)pid);
 	tdb_delete(wcache->tdb, string_tdb_data(key_str));
 
-	fstr_sprintf(key_str, "DE/%d", pid);
+	fstr_sprintf(key_str, "DE/%d", (int)pid);
 	tdb_delete(wcache->tdb, string_tdb_data(key_str));
 
 	return;
